python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hiitl"
version = "0.1.0"
description = "The control point for software that can act. Observe, enforce, and audit every action."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "HIITL, Inc."},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
    "Typing :: Typed",
]
keywords = ["ai", "agents", "policy", "control", "execution", "safety", "audit"]
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://hiitl.com"
Documentation = "https://docs.hiitl.com"
Repository = "https://github.com/hiitlhq/hiitl"
Issues = "https://github.com/hiitlhq/hiitl/issues"
Changelog = "https://github.com/hiitlhq/hiitl/releases"

[tool.setuptools.packages.find]
include = ["hiitl*"]
